requirements.txt
# Core ML/DL
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0  # YOLOv8
transformers>=4.30.0
pillow>=9.5.0
opencv-python>=4.8.0
albumentations>=1.3.0

# API & Web
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6
pydantic>=2.0.0
httpx>=0.24.0

# MLOps
mlflow>=2.5.0
dvc>=3.0.0
wandb>=0.15.0
optuna>=3.2.0

# Monitoring
prometheus-client>=0.17.0
evidently>=0.4.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0
pyyaml>=6.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
locust>=2.15.0

# Development
black>=23.7.0
flake8>=6.0.0
mypy>=1.4.0
pre-commit>=3.3.0
